requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paho-mqtt==2.1.0
requests==2.32.3
tensorflow==2.17.0
pandas==2.2.3
numpy==1.23.5
flask==3.0.3